Remove 'CHART_VERSION' variable from job
Removes the "CHART_VERSION" variable from the functional test job. This
was a testing entry from an MR that should have been deleted before
merge.

The Operator will find the latest version within the image if a
CHART_VERSION is not specified. In this case, we're specifying a chart
version that does not exist in the image.