Pin available Chart versions
Rather than downloading the latest 3 minor versions, specify the exact
versions we want to include.

This gives us the ability to pin the version of the Operator to
compatible versions of the Helm chart.

It also ensures consistent behavior regardless of when the Operator
image is built, since images built at different times could end up with
different Chart versions.