Remove 'make install_required_operators' from docs
Removes 'make install_required_operators' step from installation docs.
This is primarily a CI-specific step, and with
https://gitlab.com/gitlab-org/cloud-native/gitlab-operator/-/merge_requests/172
this step requires an environment variable to be configured that users
would not need.

Rather than using this make target, end users will continue to use the
documentation near the top of the page that outlines the dependencies
and provides links to documentation on how to install them.