Add adapter hash suffix to Migrations job name
Adds the CustomResourceAdapter's hash as a suffix to the Migrations Job
name (the first 3 characters of the hash).

I couldn't use a random hash because that value would be different every
time the controller did a reconcile loop, and would have resulted in a
brand new Migrations job deployed upon every loop.

Instead, we use the Adapter hash because that value will change every
time the CR values change, which is ultimately what we want anyway.

Note that this is expected to be a temporary workaround. The more
complete fix is being discussed in
https://gitlab.com/gitlab-org/charts/gitlab/-/issues/734.