Add `jobSucceeded` method to check Job status
Adds a method called `jobSucceeded` that will:
- Return an error if it can't find the Job in the cluster
- Return an error if the Job failed
- Return an error if the Job is still running
- Return nil if the Job succeeded

The error from the method should be passed to the caller, which can
call one of the available `requeue` functions so the controller will
try the Job until it passes. This is added as an improvement to the
previous implementation, which would wait (via a `for` loop) until
the Job passed - this meant that the controller was effectively asleep,
not responding to any updates to the CR during the wait period.

Changelog: performance