Replace the old adapter with the new one
It also contains adjustments to Helm template utility to ensure
that it works with the new adapter.

The most important change in the new adapter that impacts the
Helm template is the way that it calculates the hash based on
the UID and Generation of the underlying object. This changed
the template cache and in fact made it simpler.

The other change is that the adapter now can return its related
Chart so the template engine doesn't have to locate anymore.

The metrics resource generators are also changed to use the
adapter instead of the GitLab resource instead to make them more
generic.

Changelog: changed