Ensure "Running" phase only set if Condition true
It's possible to set the Condition "Running" to "false".
Because of this, when `test.sh` checks if "Phase" == "Running",
it concludes that the entire application is running - even if
"Condition == Running" is set to False, not True.

This updates the check to ensure that both:
- The condition type == the "running" condition
- The "running" condition is set to "true"

If both of these criteria are met, then it will set the
Status Phase to Running.

Changelog: fixed