Fail early if Chart catalog cannot be populated
Returns a failure when attempting to populate the Chart catalog.
If it's empty, we return an error that no charts could be found.

This is to address some vague errors we've seen when testing locally,
in a scenario where the Helm Charts may not have been retrieved using
./scripts/retrieve_gitlab_charts.sh.

Additionally, it's best not to ignore errors anyways. Even though the
final Operator image has the charts built in, there are always edge
cases that should be surfaced without getting a vague "nil object"
error.

Changelog: fixed