Use 'General Change.md' as default MR description
Renames 'General Change.md' to 'Default.md' so it is automatically
used as the default merge request description.

Related documentation:
https://docs.gitlab.com/ee/user/project/description_templates.html#supported-variables-in-merge-request-templates

This ensures that the '%{first_multiline_commit}' variable works as
expected, given the note in the documentation:

> This feature is available only for the default template.

The [default template docs](https://docs.gitlab.com/ee/user/project/description_templates.html#set-a-default-template-for-merge-requests-and-issues)
state that the file needs to be named 'Default.md' (case insensitive).