CI: move check_docs_links to gitlab-org tag
Move runner tags for `check_docs_links` to `gitlab-org`, which is present
on GitLab.com and Dev.

Change the CI job tags for check_docs_links job to gitlab-org, for the sake of
confirmation. Consider, in the future, removing this job from Dev entirely.
gitlab-org may not be the best tag! We do want this to run for a fork as well!

Stems from a recent [Docs CI MR](https://gitlab.com/gitlab-org/cloud-native/gitlab-operator/-/merge_requests/602).

Applies the same fix as applied in Charts:
https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/3024