Add utility for discovering managed objects
This MR introduces `kube.DiscoverManagedObjects`, a utility function that finds all Kubernetes objects that are owned by another resource. The owner is generally a CustomResource. One common usage is for a controller to discover its managed resources.

The current implementation relies on the [OwnerReferences](https://kubernetes.io/docs/concepts/overview/working-with-objects/owners-dependents/#owner-references-in-object-specifications) of objects to determine if it is managed by another object. It selects the objects that refer to the specified owner.

Related to #889

Changelog: added