Use createOrPatch for Issuer, not createOrUpdate
As reported in
https://gitlab.com/gitlab-org/cloud-native/gitlab-operator/-/issues/614,
the Issuer object was reportedly being 'updated' in every reconcile loop
despite no changes being made.

Here we use 'createOrPatch' on the Issuer instead of 'createOrUpdate'.
This was the only place where 'createOrUpdate' was used, and was most
likely a remnant due to unexpected behavior from 'createOrPatch'
historically that appears to have been addressed. Using this function,
there are no more logs of the Issuer being updated on each reconcile
loop.

Closes https://gitlab.com/gitlab-org/cloud-native/gitlab-operator/-/issues/614

Changelog: fixed