Revert Set global.hosts.externalIP on NGINX Service
Reverts
https://gitlab.com/gitlab-org/cloud-native/gitlab-operator/-/merge_requests/328/diffs?commit_id=a576cf2777238e754a0789821300cede8a983671

Context from commit message:

>>>
This restores original functionality that was mistakenly removed in
https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/2267.

It's been fixed in the Chart itself in
https://gitlab.com/gitlab-org/charts/gitlab/-/merge_requests/2317,
but is not part of Chart version 5.6.0.

This can be removed when Chart version 5.6.1 is added.
>>>

Now that all versions in ./CHART_VERSIONS are above the 5.6.1,
this workaround can be removed.

Changelog: other