Delete unused resources
Catalogs the existing and desired managed resources and at the end
of the reconciliation loop deletes the unused resources (those that
exist but not desired).

This is done via an interface and helper functions to minimize the
impact on the existing controller code. This is not ideal but they
will be removed once we could replace it with the framework.

Changelog: fixed