Remove opm from .tool-versions
The opm binary is not called directly. The test_olm.sh script
downloads it at a specified version.

Additionally, this binary is managed in a custom repository and the
syntax is not supported by alternative version managers such as
https://github.com/jdxcode/rtx, which is mostly asdf compatible
except for the syntax with the git URL after the version.